+++ apps/api/.dockerignore
@@ -1,0 +1,5 @@
+tests/
+
+**/__pycache__/
+
+.*
