+++ apps/api/.env.example
@@ -1,2 +1,3 @@
 FLASK_ENV=development
-API_KEY=dummy_key_for_development
+API_KEY=your_development_api_key
+ALLOWED_ORIGINS=http://localhost:5173,https://yourdomain.com
