+++ apps/api/requirements.txt
@@ -1,4 +1,6 @@
 Flask==3.1.0
-pytest==8.3.4
-pytest-flask==1.3.0
 python-dotenv==1.0.1
+palettum==0.3.2
+gunicorn==21.2.0
+python-dotenv==1.0.1
+flask-cors==4.0.0
