+++ apps/frontend/Dockerfile
@@ -1,4 +1,4 @@
-FROM oven/bun:1
+FROM oven/bun:1 as builder
 
 WORKDIR /app
 
@@ -7,6 +7,17 @@ RUN bun install
 
 COPY . .
 
-ENV VITE_API_URL=http://localhost:5000
+ARG VITE_API_URL
+ENV VITE_API_URL=$VITE_API_URL
 
-CMD ["bun", "run", "dev", "--host"]
+RUN bun run build
+
+FROM nginx:alpine
+
+COPY --from=builder /app/dist /usr/share/nginx/html
+# Add Nginx configuration to handle SPA routing
+COPY nginx.conf /etc/nginx/conf.d/default.conf
+
+EXPOSE 80
+
+CMD ["nginx", "-g", "daemon off;"]
