+++ apps/frontend/src/App.tsx
@@ -3,7 +3,7 @@ import ImageUpload from "@/components/ImageUpload";
 import ImageDimensions from "@/components/ImageDimensions";
 import PaletteManager from "@/components/PaletteManager";
 import PalettifyImage from "@/components/PalettifyImage";
-import type { PaletteColor } from "@/services/api";
+import type { Color } from "@/lib/palettes/types";
 import ImageTransparency from "@/components/ImageTransparency";
 
 function App() {
@@ -12,7 +12,7 @@ function App() {
     width: null as number | null,
     height: null as number | null,
   });
-  const [selectedPalette, setSelectedPalette] = useState<PaletteColor[]>([]);
+  const [selectedPalette, setSelectedPalette] = useState<Color[]>([]);
   const [transparentThreshold, setTransparentThreshold] = useState<number>(0);
 
   const handleFileSelect = useCallback((file: File | null) => {
@@ -29,7 +29,7 @@ function App() {
   const handleThresholdChange = useCallback((newThreshold: number) => {
     setTransparentThreshold(newThreshold);
   }, []);
-  const handlePaletteSelect = useCallback((colors: PaletteColor[]) => {
+  const handlePaletteSelect = useCallback((colors: Color[]) => {
     setSelectedPalette(colors);
   }, []);
