--- docker-compose.yml
@@ -1,32 +1,0 @@
-version: "3.8"
-services:
-  api:
-    build:
-      context: ./apps
-      dockerfile: api/Dockerfile
-    ports:
-      - "5001:5000"
-    volumes:
-      - ./apps/api:/app/api
-      - ./apps/core:/app/core
-    environment:
-      - FLASK_ENV=development
-      - API_KEY=${API_KEY}
-      - PYTHONPATH=/app/core:/app
-
-  frontend:
-    build:
-      context: ./apps/frontend
-    ports:
-      - "5173:5173"
-    volumes:
-      - ./apps/frontend:/app
-      - /app/node_modules
-    environment:
-      - DOCKER_ENV=true
-    depends_on:
-      - api
-
-networks:
-  default:
-    driver: bridge
